+++ src/math.rs
@@ -1,10 +1,10 @@
 #![forbid(unsafe_code)]
 
-pub fn is_prime_number(number: u128) -> bool {
+pub fn is_prime_number(number: i32) -> bool {
     if number % 2 == 0 {
         return false;
     }
-    let mut counter: u128 = 3;
+    let mut counter: i32 = 3;
     while counter < number {
         if number % counter == 0 {
             return false;
