+++ src/user_input.rs
@@ -2,14 +2,14 @@
 
 use std::io;
 
-pub fn get_user_input(prompt: &str) -> u128 {
+pub fn get_user_input(prompt: &str) -> i32 {
     println!("{}", prompt);
 
     let mut user_input: String = String::new();
 
     io::stdin().read_line(&mut user_input).expect("Failed to read line");
 
-    let number: u128 = user_input.trim().parse().expect("Please enter an integer!");
+    let number: i32 = user_input.trim().parse().expect("Please enter an integer!");
 
     return number;
 }
